# interface/pkg/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "認証情報から自分のユーザー情報を取得する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentiation header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Something wrong",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Something wrong",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something wrong",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "認証情報とリクエスト情報からユーザーを作成する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentiation header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.userCreateReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "500": {
                        "description": "Something wrong",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.userCreateReq": {
            "type": "object",
            "required": [
                "nickname"
            ],
            "properties": {
                "nickname": {
                    "type": "string"
                }
            }
        }
    }
}